Corregido error 403 en catálogo y sincronizados de IDs con el Backend

+++ app/src/main/java/com/example/desarrollo/model/CartItemDetails.kt
@@ -1,11 +1,7 @@
 package com.example.desarrollo.model
 
-/**
- * Clase de datos que une la información de un Producto con la cantidad en el carrito.
- * Room usará esta clase para devolver el resultado de una consulta JOIN.
- */
 data class CartItemDetails(
-    val productId: Int,
+    val productId: Long, // 🎯 Cambiado a Long
     val name: String,
     val price: Int,
     val imageRes: Int,
